+++ spec/unit/mongoid/config/environment_spec.rb
@@ -59,8 +59,10 @@ describe Mongoid::Config::Environment do
 
     context "when no environment information is found" do
 
-      it "returns the default mongoid environment" do
-        described_class.env_name.should eq("development")
+      it "raises an error" do
+        expect { described_class.env_name }.to raise_error(
+          Mongoid::Errors::NoEnvironment
+        )
       end
     end
   end